fix tests & deps
