+++ package.json
@@ -30,12 +30,12 @@
   "devDependencies": {
     "@babel/eslint-parser": "^7.23.3",
     "@biomejs/biome": "1.5.2",
-    "@diary-spo/shared": "1.8.8",
+    "@diary-spo/shared": "1.8.9",
     "@diary-spo/types": "latest",
     "@preact/preset-vite": "^2.8.1",
     "@rollup/plugin-terser": "^0.4.4",
     "@types/node": "^20.11.5",
-    "@vitest/coverage-v8": "^1.1.3",
+    "@vitest/coverage-v8": "^1.2.1",
     "@vkontakte/vk-miniapps-deploy": "^0.1.6",
     "gh-pages": "^6.0.0",
     "jsdom": "^23.2.0",
