+++ src/utils/logOut/__tests__/i.test.ts
@@ -1,0 +1,8 @@
+import { logOut } from '@utils'
+import { expect, it } from 'vitest'
+
+/** logOut **/
+it('должна очистить локальное хранилище', () => {
+  logOut()
+  expect(localStorage.length).toBe(0)
+})
