+++ src/utils/textToLink/__test__/mocs.ts
@@ -1,0 +1,2 @@
+export const mockTextWithLinks =
+  'Check out this website: https://www.example.com.'
